requirements.txt
alembic==1.14.1
annotated-types==0.7.0
anyio==4.8.0
asyncpg==0.30.0
click==8.1.8
fastapi==0.115.8
GeoAlchemy2==0.17.0
greenlet==3.1.1
h11==0.14.0
idna==3.10
Mako==1.3.9
MarkupSafe==3.0.2
numpy==2.2.3
packaging==24.2
psycopg==3.2.4
pydantic==2.10.6
pydantic_core==2.27.2
shapely==2.0.7
sniffio==1.3.1
SQLAlchemy==2.0.38
starlette==0.45.3
typing_extensions==4.12.2
uvicorn==0.34.0
python-dotenv~=1.0.1